+++ src/_utils/mediaSizeProvider/index.unit.tsx
@@ -1,9 +1,15 @@
 import React, { useContext } from 'react'
 import { mount } from 'enzyme'
 
-import { MediaSizeContext, MediaSizeProvider } from './index'
+import {
+  MediaSize,
+  MediaSizeContext,
+  MediaSizeProvider,
+  useIsLargeMediaSize,
+  useIsSmallMediaSize,
+} from './index'
 
-const mockWindowInnerWidth = value => {
+const mockWindowInnerWidth = (value: number): void => {
   Object.defineProperty(window, 'innerWidth', {
     writable: true,
     configurable: true,
@@ -11,12 +17,19 @@ const mockWindowInnerWidth = value => {
   })
 }
 
-let ChildComponent
 describe('MediaSizeProvider', () => {
+  let ChildComponent: any
+
   beforeEach(() => {
     ChildComponent = () => {
-      const value = useContext(MediaSizeContext)
-      return <div>{value}</div>
+      const valueFromContext = useContext(MediaSizeContext)
+      const isSmall = useIsSmallMediaSize().toString()
+      const isLarge = useIsLargeMediaSize().toString()
+      return (
+        <div>
+          context: {valueFromContext} | isSmall: {isSmall} | isLarge: {isLarge}
+        </div>
+      )
     }
   })
   it('Should provide "small" value when screen is under `responsiveBreakpoints.small`', () => {
@@ -27,7 +40,8 @@ describe('MediaSizeProvider', () => {
       </MediaSizeProvider>,
     )
 
-    expect(wrapper.find('div').text()).toEqual('small')
+    const expected = 'context: small | isSmall: true | isLarge: false'
+    expect(wrapper.find('div').text()).toEqual(expected)
   })
 
   it('Should provide "large" value when screen is above `responsiveBreakpoints.small`', () => {
@@ -38,6 +52,27 @@ describe('MediaSizeProvider', () => {
       </MediaSizeProvider>,
     )
 
-    expect(wrapper.find('div').text()).toEqual('large')
+    const expected = 'context: large | isSmall: false | isLarge: true'
+    expect(wrapper.find('div').text()).toEqual(expected)
+  })
+
+  it('Should provide "small" when no provider', () => {
+    mockWindowInnerWidth(900)
+    const wrapper = mount(<ChildComponent />)
+
+    const expected = 'context: small | isSmall: true | isLarge: false'
+    expect(wrapper.find('div').text()).toEqual(expected)
+  })
+
+  it('Should respect the mediaSizeForTestsOnly override`', () => {
+    mockWindowInnerWidth(300)
+    const wrapper = mount(
+      <MediaSizeProvider mediaSizeForTestsOnly={MediaSize.LARGE}>
+        <ChildComponent />
+      </MediaSizeProvider>,
+    )
+
+    const expected = 'context: large | isSmall: false | isLarge: true'
+    expect(wrapper.find('div').text()).toEqual(expected)
   })
 })